requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
python-decouple==3.8
pandas==2.3.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
gunicorn==23.0.0
whitenoise==6.9.0

# For data processing
numpy>=2.3.0

# For development
django-debug-toolbar==4.4.6

# For plotting and charts (optional, for statistics)
plotly==5.24.1
matplotlib>=3.9.0 